Raise error on invalid URL
Previuosly, valid_url? returned false if the URL given couldn't be reached. This resulted in Vanagon interpreting this as not a URL at all, and then giving a very confusing error. Instead, when the URL itself is in a valid format but we can't actually reach it, we raise an exception saying so instead.