Add source_date_epoch to project object
This creates the source_date_epoch variable on the project object, picking it up from the SOURCE_DATE_EPOCH env var, or using the current time if that isn't set. It converts all times to UTC so it doesn't matter what time zone you are in when you run it.