Skip the `dh_perl` action
The openvox-agent pacakage contains some bits from openssl, in
particular a perl script `tsget.pl` and a symbolic link `tsget` that
points to it.

`dh_perl` wants to normalize shebangs in perl files, but depending on the
order files are enumerated, it breaks builds reproducibility:
  * When `dh_perl.pl` is found first, it is normalized (replaced by a new
    regular file with the normalized shebang), and then when
    `dh_perl` is found it does not need normalizing and is kept
    unchanged.
    The package has `dh_perl.pl` as a regular file, and `dh_perl` as a
    symbolic link to that file.
  * When `dh_perl` is found first, it is normalized (replaced by a new
    regular file with the normalized shebang), and then `dh_perl.pl` is
    found and it is also normalized.
    The package has `dh_perl.pl` and `dh_perl` as two regular files.

While some debhelpers can have support for an `-X` flag to exclude
files, `dh_perl` ignore it, so we cannot rely on it.

Because perl files in the OpenVox packages are only third-party tools,
we can skip this helper to avoid breaking builds reproducibility.