Fix reproducibitily issue with user-defined abbrev length
Users may tune the length of commit IDs abbreviation with something
like:

```
git config --global core.abbrev 12
```

This is reflected in the version / filename of the packages when not
building a tag.

By default, git use an abbreviation length that depends of the number
of objects in the repo.  For the OpenVoxProject/openvox-agent project,
that is currently 9, so hardcode this value so that it does not depend
of users local configuration.

Fixes #3